Internal changes to allow over 2,147,483,647 blocks to be queued for drawing simultaneously on a map.

+++ branch-0.60x/fCraft/World/Map.cs
@@ -339,10 +339,10 @@ namespace fCraft {
             get { return drawOps.Count; }
         }
 
-        public int DrawQueueBlockCount {
+        public long DrawQueueBlockCount {
             get {
                 lock( drawOpLock ) {
-                    return drawOps.Sum( op => op.BlocksLeftToProcess );
+                    return drawOps.Sum( op => (long)op.BlocksLeftToProcess );
                 }
             }
         }
